+++ src/test/java/ch/bitagent/bitcoin/lib/wallet/WalletTest.java
@@ -33,7 +33,7 @@ class WalletTest {
     void privkey() {
         var extendedPrivkey = ExtendedKey.parse("zprvAdG4iTXWBoARxkkzNpNh8r6Qag3irQB8PzEMkAFeTRXxHpbF9z4QgEvBRmfvqWvGp42t42nvgGpNgYSJA9iefm1yYNZKEm7z6qUWCroSQnE");
         var wallet = Wallet.parse(extendedPrivkey);
-        assertEquals(20, wallet.getAddressList0().size());
+        assertEquals(10, wallet.getAddressList0().size());
         assertEquals(10, wallet.getAddressList1().size());
         wallet.history();
         log.info(wallet.toString());
@@ -43,7 +43,7 @@ class WalletTest {
     void pubkey() {
         var extendedPubkey = ExtendedKey.parse("zpub6rFR7y4Q2AijBEqTUquhVz398htDFrtymD9xYYfG1m4wAcvPhXNfE3EfH1r1ADqtfSdVCToUG868RvUUkgDKf31mGDtKsAYz2oz2AGutZYs");
         var wallet = Wallet.parse(extendedPubkey);
-        assertEquals(20, wallet.getAddressList0().size());
+        assertEquals(10, wallet.getAddressList0().size());
         assertEquals(10, wallet.getAddressList1().size());
     }
 
@@ -51,7 +51,7 @@ class WalletTest {
     void mnemonicSentenceSignVerify() {
         var mnemonicSentence = Properties.getWalletMnemonic(Properties.WALLET_FILENAME, 1);
         var wallet = Wallet.parse(mnemonicSentence, null);
-        assertEquals(20, wallet.getAddressList0().size());
+        assertEquals(10, wallet.getAddressList0().size());
         assertEquals(10, wallet.getAddressList1().size());
         var address = wallet.getAddressList0().get(0).getAddressString();
         var message = "Test";